Updated ship object to have current/max/base hp as well as base/current stats
Updeted ship object to have base and modified stats
Broke ship objects items into crew & ship items
Add and remove from ship parameters by adding items and crew
Ship and hull HP are stored as tuples (Max, Current)

+++ DontSink/Assets/Scripts/Objects/ItemObjects/SailObject.cs
@@ -4,8 +4,8 @@ using System.Collections;
 public class SailObject : ItemObject
 {
     //Fields
-    private double speed;
+    private int speed;
 
     //Accessors
-    public double Speed { get { return speed; } set { speed = value; } }
+    public int Speed { get { return speed; } set { speed = value; } }
 }
